+++ R/ggKAP.R
@@ -28,12 +28,13 @@
 #' @param show_pval a logical, should the p-value of a log-rank test appear on the plot? Default is `TRUE` if `data` alone is provided, and there are several Kaplan-Meier curves in each panel of the plot.
 #' @param show_median a logical, should the median survival (and associated legend) appear on the plot? Default is `TRUE` if `data` alone is provided, `FALSE` if `sdata` is also provided.
 #' @param show_sim a logical, should a confidence interval (and associated legend) computed from simulated data (`sdata`) appear on the plot? Default is `TRUE` if `sdata` is provided.
-#' @param ci_level a numeric, between 0 and 1, indicating the level of the confidence interval. Default is `.95` (95%) if the confidence interval is calculated from observed data (`data`), `.90` (90%) if calculated from simulations (`sdata`).
 #' @param cuminc a logical, should the plot represent the cumulative incidence? If `FALSE`, the default, the survival is represented.
+#' @param ci_level a numeric, between 0 and 1, indicating the level of the confidence interval. Default is `.95` (95%) if the confidence interval is calculated from observed data (`data`), `.90` (90%) if calculated from simulations (`sdata`).
 #' @param ci_alpha a numeric, between 0 and 1, indicating the opacity of the confidence interval.
 #' @param censor_shape shape of the censoring event.
 #' @param censor_size a numeric, the size of the censoring event.
 #' @param median_linetype a numeric, linetype of the median survival line.
+#' @param pval_pos a vector of 2 numeric, coordinate positions of the p-value text (`c(x,y)`). Default in `NULL`, should appear on the bottom left.
 #' @param scale_x_break_by a numeric, interval width when the x-axis (i.e. time) will be broken (e.g. every `4` weeks), as well as the times when the number of patient at risk is calculated. Default is `NULL` (calls the default setup of ggplot2).
 #' @param scale_y_labels passed to `ggplot2::scale_y_continuous(labels = )` when the Kaplan-Meier curve is constructed. Default is `scales::percent` to label the survival as percentage. Set to `identity` or `ggplot2::waiver()` for the original numeric values.
 #' @param label_x a character, the name of the x-axis (i.e. time) of the Kaplan-Meier figure and of the risk table.
@@ -48,8 +49,8 @@
 #' @param scale_color_labels passed to `ggplot2::scale_color_manual(labels = )` for the Kaplan-Meier figure, and `ggplot2::scale_y_discrete(labels = )` for the risk table. The default is `waiver()` (calls the default setup of ggplot2), but can accept `scales::label_parsed` if the coloring variable is an expression (`expression()`) and special character or super/subscripts should be shown.
 #' @param scale_color_values passed to `ggplot2::scale_color_manual(values = )` for the Kaplan-Meier figure and the risk table. Default will use the default PMX graphic charter, but named vectors of colors (with names matching with factor levels) are welcome. Levels not available in the data will be dropped.
 #' @param facetting_args a list, arguments that will overwrite the default arguments of `facet_wrap()`. Note that it will not overwrite `vars`, the latter being informed by `facet_var`. Use `list(labeller = ggplot2::label_parsed)` if the facetting variable is an expression (`expression()`) and special character or super/subscripts should be shown.
-#' @param xlim a vector of 2 numeric values, limits of the x-axes. Default will use the default ggplot setup.
-#' @param ylim a vector of 2 numeric values, limits of the y-axis of the Kaplan-Meier figure. Default is 0 to 1.
+#' @param xlim a vector of 2 numeric values, limits of the x-axes of the Kaplan-Meier plot and risk table. Default will use the default ggplot setup. Passed to `ggplot2::coord_cartesian()`.
+#' @param ylim a vector of 2 numeric values, limits of the y-axis of the Kaplan-Meier figure. Default is 0 to 1. Passed to `ggplot2::coord_cartesian()`.
 #' @param title_risktable a character, the title above the risk table.
 #' @param arrange a logical, should the elements of the plot be arranged in a single plot with `cowplot::grid.arrange()` (default is `TRUE`). If `FALSE`, a list with separated elements will be returned (useful for advanced customization.)
 #' @param rel_heights a list specifying the relative heights (1) between the legends (default is 1 and 1, same height), (2) between the Kaplan-Meier figure and the risk table (default is 2 and 1, i.e. figure twice higher than table) and (3) overall (default is 1 and 8, combined legends takes 1/9 of heights, combined figure and tables takes 8/9 of heights)
@@ -92,6 +93,31 @@
 #'
 #' ggKAP(dat, scale_x_break_by = 12) #Both risk table and graph are updated
 #'
+#' ggKAP(
+#'   dat, color_var = "SEXF",
+#'   xlim = c(0, 50), ylim = c(.10, .5)
+#' ) # limits do not change the underlying data (e.g.: same p value)
+#'
+#' ggKAP(
+#'   dat, color_var = "SEXF",
+#'   pval_pos = c(80, .8)
+#' ) # manually define p value position on the top right
+#'
+#' # Labelling
+#'
+#' ggKAP(dat, color_var = "AUCQF")
+#'
+#' # This will update both the "color legend" label and the y axis of the risk table
+#' ggKAP(dat, color_var = "AUCQF", label_color = "Nice color label")
+#'
+#' # But it can be differentiated if needed
+#' ggKAP(dat, color_var = "AUCQF",
+#'   label_color = "Nice color label",
+#'   label_y_risktab = "Label for the\ny-axis of risk table"
+#'   )
+#'
+#'
+#'
 #' ggKAP(dat, rel_heights = list(
 #'   legends = c(1,1),
 #'   figtable = c(8,2),
@@ -105,6 +131,8 @@
 #'
 #' ggKAP(dat, simdat, facet_var = "SEXF")
 #' ggKAP(sdata = simdat, facet_var = "SEXF")
+#'
+#' # Special characters
 #'
 #' ggKAP(dat, color_var = "AUCQF")
 #' ggKAP(dat, color_var = "AUCQF", label_color = expression(AUC['0-24h']))
@@ -127,12 +155,13 @@ ggKAP <- function(data,
                   show_pval = !missing(data)&&missing(sdata)&&!is.null(color_var)&&(if(!is.null(facet_var)){color_var!=facet_var}else{TRUE}),
                   show_median = !missing(data)&&missing(sdata),
                   show_sim = !missing(sdata),
-                  ci_level = if(show_se) .95 else .90,
                   cuminc = FALSE,
+                  ci_level = if(show_se) .95 else .90,
                   ci_alpha = .5,
                   censor_shape = "|",
                   censor_size = 4,
                   median_linetype = 2,
+                  pval_pos = NULL,
                   scale_x_break_by = NULL,
                   scale_y_labels = scales::percent,
                   label_x = "Time",
@@ -242,7 +271,9 @@ ggKAP <- function(data,
     fig <- fig +
       stat_kaplanmeier_pval(
         aes(strata = eval(..col)),
-        data = data
+        data = data,
+        xpos = pval_pos[1],
+        ypos = pval_pos[2]
       )
   }
 
@@ -327,11 +358,9 @@ ggKAP <- function(data,
   fig <- fig +
     scale_x_continuous(
       breaks = scale_x_breaks,
-      limits = xlim
     ) +
     scale_y_continuous(
       labels = scale_y_labels,
-      limits = ylim
     ) +
     scale_color_manual(
       labels = scale_color_labels,
@@ -343,13 +372,24 @@ ggKAP <- function(data,
   risktab <- risktab +
     scale_x_continuous(
       breaks = scale_x_breaks,
-      limits = xlim
     ) +
     scale_y_discrete(
       labels = scale_color_labels
     ) +
     scale_color_manual(
       values = scale_color_values
+    )
+
+  # Coordinates
+  fig <- fig +
+    coord_cartesian(
+      xlim = xlim,
+      ylim = ylim
+    )
+
+  risktab <- risktab +
+    coord_cartesian(
+      xlim = xlim,
     )
 
   # Guides
@@ -359,6 +399,7 @@ ggKAP <- function(data,
       fill = guide_legend(order = 2),
       shape = guide_legend(order = 3)
     )
+
 
   # Titles
   # labs() don't accept NULL as input, so need to wrap with an if()
@@ -377,7 +418,7 @@ ggKAP <- function(data,
       labs(
         subtitle = title_risktable,
         x = label_x,
-        y = label_color
+        y = label_y_risktab
       )
   }