+++ tests/testthat/test-makeSummaryTTE.R
@@ -30,6 +30,7 @@ test_that("makeSummaryTableTTE works with RTTE", {
                               innerLevel   = "DOSEN",
                               innerLabel   = "Dose",
                               showAvnObs = TRUE,
+                              avnObsPercent = FALSE,
                               asList = TRUE)
   expect_equal(as.double(ans[[3]]$avnObs), ans[[3]]$nObs / ans[[3]]$subjects, tolerance = .001)
   ans <- makeSummaryTableTTE(rttedata,
@@ -54,6 +55,24 @@ test_that("makeSummaryTableTTE works with TTE", {
                              asList = TRUE)
   expect_equal(ans[[3]]$subjects, c(rep(200, 4), rep(150, 3)))
   expect_equal(ans[[3]]$nObs, c(108, 86, 74, 75 ,70, 68, 56))
+  expect_equal(ans[[3]]$avnObs[c(1, 4)], c("54.0%", "37.5%")) # checking formatting only
+  ans <- makeSummaryTableTTE(tt1edata,
+                             outerLevel   ="STUDYIDN" ,
+                             outerLabel   = "Study",
+                             innerLevel   = "DOSEN",
+                             innerLabel   = "Dose",
+                             asList = TRUE,
+                             digits = 4)
+  expect_equal(ans[[3]]$avnObs[c(1, 4)], c("54.00%", "37.50%")) # checking formatting only
+
+  ans <- makeSummaryTableTTE(tt1edata,
+                             outerLevel   ="STUDYIDN" ,
+                             outerLabel   = "Study",
+                             innerLevel   = "DOSEN",
+                             innerLabel   = "Dose",
+                             asList = TRUE,
+                             avnObsPercent = FALSE)
+  # checking numbers
   expect_equal(as.double(ans[[3]]$avnObs), ans[[3]]$nObs / ans[[3]]$subjects, tolerance = .001)
 
 
@@ -66,3 +85,5 @@ test_that("makeSummaryTableTTE works with TTE", {
     capture.output()
   expect_true(stringr::str_detect(ans[27], "\\{c\\}Proportion of "))
 })
+
+